Update ipynb file with two forward pass options
I have added a second forward pass option that is more suitable for this case. This is an option that implements a method called fastforward in the RNN class. What it does is the typical forward pass but much faster than the original option. This is because only the last element of the sequence is passed to the fc-layer at the output of the LSTM.